Extract capabilities to dedicated files
