+++ tests/test_capability_toggle.py
@@ -5,7 +5,7 @@ import pytest
 from pytest_homeassistant_custom_component.common import async_mock_service
 
 from custom_components.yandex_smart_home import const
-from custom_components.yandex_smart_home.capability import CAPABILITIES_TOGGLE
+from custom_components.yandex_smart_home.capability_toggle import CAPABILITIES_TOGGLE
 from custom_components.yandex_smart_home.const import (
     TOGGLE_INSTANCE_MUTE,
     TOGGLE_INSTANCE_OSCILLATION,
